Reading environment variable from C
