+++ Sources/AWSSDKSwiftCore/Doc/Environment.swift
@@ -1,0 +1,14 @@
+#if os(Linux)
+import Glibc
+#else
+import Darwin.C
+#endif
+
+internal enum Environment {
+    internal static subscript(_ name: String) -> String? {
+        guard let value = getenv(name) else {
+            return nil
+        }
+        return String(cString: value)
+    }
+}
