+++ Sources/AWSSignerV4/credentials.swift
@@ -12,8 +12,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-import class Foundation.ProcessInfo
-
 /// Protocol for providing credential details for accessing AWS services
 public protocol Credential {
     var accessKeyId: String {get}
@@ -33,23 +31,3 @@ public struct StaticCredential: Credential {
         self.sessionToken = sessionToken
     }
 }
-
-/// environment variable version of credential that uses system environment variables to get credential details
-public struct EnvironmentCredential: Credential {
-    public let accessKeyId: String
-    public let secretAccessKey: String
-    public let sessionToken: String?
-    
-    public init?() {
-        guard let accessKeyId = ProcessInfo.processInfo.environment["AWS_ACCESS_KEY_ID"] else {
-            return nil
-        }
-        guard let secretAccessKey = ProcessInfo.processInfo.environment["AWS_SECRET_ACCESS_KEY"] else {
-            return nil
-        }
-        self.accessKeyId = accessKeyId
-        self.secretAccessKey = secretAccessKey
-        self.sessionToken = ProcessInfo.processInfo.environment["AWS_SESSION_TOKEN"]
-    }
-}
-
